installed rust on my desktop
if there are weird things that happen after this, then that is why. I just installed rust onto my gaming desktop where i will potentially be coding a game with the bevy engine. I am committing this so I will not have like 700 changes waiting to be commited on my machine.